+++ src/main/java/dansplugins/dpm/commands/HelpCommand.java
@@ -18,7 +18,7 @@ public class HelpCommand extends AbstractPluginCommand {
         commandSender.sendMessage(ChatColor.AQUA + "=== DPM Commands ===");
         commandSender.sendMessage(ChatColor.AQUA + "/dpm help - View this list of commands.");
         commandSender.sendMessage(ChatColor.AQUA + "/dpm list [installed|available] - List plugins, optionally filtered.");
-        commandSender.sendMessage(ChatColor.AQUA + "/dpm get <plugin-name> [plugin-name ...] - Download one or more plugins.");
+        commandSender.sendMessage(ChatColor.AQUA + "/dpm get <plugin-name> [plugin-name ...] [--experimental|--stable] - Download one or more plugins, optionally switching release channel.");
         commandSender.sendMessage(ChatColor.AQUA + "/dpm clean [--confirm] - Preview or remove duplicate plugin JARs.");
         commandSender.sendMessage(ChatColor.AQUA + "/dpm stats - View plugin statistics.");
         commandSender.sendMessage(ChatColor.AQUA + "/dpm update [plugin-name ...] - Update all installed plugins, or specific ones.");